+++ packages/webapp/src/api/api.ts
@@ -8,6 +8,10 @@ const mapEntriesForBrowser = entry => {
     const ascii = decodeBase64(entry.similarityHash);
     entry.similarityHash = ascii;
   }
+  // reset textCache from polluted fetch command <= 1.4.1
+  if (entry.textCache) {
+    entry.textCache = false
+  }
   return entry;
 }